# [browser][non-icu] `HybridGlobalization` faster encoding for change case.