Implements a chunk of web-api based globalization. Is a part of HybridGlobalization feature and contributes to https://github.com/dotnet/runtime/issues/79989.

Speeds up the decoding process:
| **measurement** | **time without this PR [ms]** | **time with this PR [ms]** |  **time ICU4C [ms]** | **increase vs ICU [times]** |
|-:|-:|-:|-:|-:|
|               String, TextInfo ToLower |     2.4686 |     1.6429 |   0.4260 |  3,80  |
|               String, TextInfo ToUpper |     2.5142 |     1.0003 |    0.4432 |   2,25   |
|            String, TextInfo ToTileCase |     47.0515 |    44.8951 | 14.8432 |  3,02    |

An attempt of encoding implementation using `TextEncoder` was actually slower than existing one, so it did not get into this PR.
| measurement |**time with this PR [ms]** |**time with this PR + `TextEncoder` [ms]**
|-:|-:|-:|
|               String, TextInfo ToLower |      1.6429ms |     3.0513ms |
|               String, TextInfo ToUpper |       1.0003ms |     3.1530ms |
|            String, TextInfo ToTileCase |      44.8951ms |    55.6864ms |


Relevant discussions:

invariant change case on C# side: https://github.com/dotnet/runtime/pull/85510#discussion_r1180385539
TextEncoder usage implications: https://github.com/dotnet/runtime/pull/85510#discussion_r1180413431